favourite library

+++ app/Http/Controllers/UnitController.php
@@ -61,19 +61,29 @@ class UnitController extends Controller
 
         //insert next unit if it does exist
         $unitt = Unit::find($id);
+        $user_unit = DB::table('unit_user')->where('unit_id', $id)->where('user_id', $user->id)->first();
+        if ($user_unit) {
+            //set old unit status to one (completed)
+            $user->units()->detach($unitt, ['status' => 0]);
+            $user->units()->attach($unitt, ['status' => 1, 'quiz_mark' => request('quiz_mark')]);
+        }
         $new_unit = Unit::where('id', '>', $unitt->id)->where('language_id', $unitt->language_id)->first();
         if ($new_unit) {
-            DB::table('unit_user')->insert([
-                'unit_id' => $new_unit->id,
-                'user_id' => $user->id,
-            ]);
+            $exist = DB::table('unit_user')->where('unit_id', $new_unit->id)->where('user_id', $user->id)->first();
+            if ($exist) {
+            } else {
+                DB::table('unit_user')->insert([
+                    'unit_id' => $new_unit->id,
+                    'user_id' => $user->id,
+                ]);
+            }
         }
 
-        //set old unit status to one (completed)
-        $user->units()->detach($unitt, ['status' => 0]);
-        $user->units()->attach($unitt, ['status' => 1]);
-
-        $user->unlock(new CompleteFirstUnit());
+        ProfileController::increasePoints(100);
+        $details = $user->achievementStatus(new CompleteFirstUnit());
+        if ($details->unlocked_at == null) {
+            $user->unlock(new CompleteFirstUnit());
+        }
 
         return response($new_unit, 200);
     }
@@ -95,6 +105,25 @@ class UnitController extends Controller
             'message' => 'Unit Created Successfully'
         ];
         return response($response, 201);
+    }
+
+    public function toggle($id)
+    {
+        $unit = Unit::find($id);
+
+        if ($unit->unit_status == 'Done') {
+            $unit->update([
+                'unit_status' => 'Undone'
+            ]);
+        } else {
+            $unit->update([
+                'unit_status' => 'Done'
+            ]);
+        }
+
+        $response = "Toggle operation done successfully";
+
+        return response($response, 200);
     }
 
     public function destroy($id)
